pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sem-merge"
version = "0.1.0"
description = "Pre-commit hook for semantic document merging using AI"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    "gitpython>=3.1.44",
    "openai>=1.82.1",
]

[project.urls]
Homepage = "https://github.com/your-org/sem-merge"
Repository = "https://github.com/your-org/sem-merge.git"
Issues = "https://github.com/your-org/sem-merge/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/sem_merge"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
    "pytest-cov>=6.1.1",
    "pyrefly>=0.17.1",
    "ruff>=0.11.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "S", "B", "A", "C4", "T20", "PT", "Q"]
ignore = ["S101", "T201"]  # Allow assert statements in tests and print statements

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
] 